[refactor] 일정 출장, 반차 내용 추가
* 일정 로직에서 출장과 반차 내용을 추가
* 반차는 0.5일로 계산

+++ src/main/java/com/rko/huemanager/domain/Employee.java
@@ -43,7 +43,7 @@ public class Employee implements UserDetails {
     private Role role;
 
     @Setter
-    private Integer vacationCount = 15;
+    private Double vacationCount = 15.0;
 
     @Column(nullable = false)
     @Setter
